Fix comment shorter than indent length in mapping
When a comment is the first item in a mapping, we used to first emit the
comment, and then the next iteration would increase the indentation and
emit the first real key in the mapping.

However, increasing the indentation does not work as you'd expect if the
current line is not empty and its length is shorter than what the
indentation would be increased to, resulting in no newline being
appended.

This commit changes the comment emitter to also emit a newline (since
comments are always followed by one).